Attempting to clean git repo
The git repo currently tracks a lot of uneeded cache files and other junk that we do not need to track. I have updated the .gitgnore, but the files also need to be explicitly deleted so we stop tracking them. This is the initial wave of cleaning to see if this cuts down on the junk.