First Attempt at shooting (STILL NEEDS TONS OF WORK)
Here, i created a projectile class. Projectiles are stored in a list in the player class and rendered when the player clicks. At the moment, i like the look of the projectile class, but we may render the projectiles somewhere else (weapon class)? Projectiles don't have direction yet but they do kind of follow the player.