In Room.cs, added conditionals to JSON object creation, and added 2D array that checks which grid tiles are occupied.
Conditionals added check to see if the barrier is of a correct size, and if the barrier fits within the grid.
The occupiedTilesGrid 2D array returns a different integer depending on if a tile is occupied wth an object. A 0 means no object is on the tile, a 1 means an object is on the tile. Currently only checks for barriers.